version 1.0.0: - Changed the PointData struct to include the time point and changed the algorithms implementation and interface accordingly, updated the FileReader class accordingly - Added CommManager class to receive incoming data using a ICommChannel and parse it + sending the results back - Implemented the MainProcces logic - Changed types for the WinTime from double to long int - Added ICommChannel interface for a communication channel - Implemented a varation of ICommChannel for local testing - LocalFileCommChannelFacade: reading the data from a file and simulating a communication channel

+++ TestedOBCApp/Timer/WinTimer.cpp
@@ -17,7 +17,7 @@ void WinTimer::startTimer()
 
     // Get the duration in microseconds
     auto duration = now.time_since_epoch();
-    double microseconds = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
+    long int microseconds = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
     m_startTime= microseconds;
 }
 
@@ -28,7 +28,7 @@ void WinTimer::stopTimer()
 
     // Get the duration in microseconds
     auto duration = now.time_since_epoch();
-    double microseconds = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
+    long int microseconds = std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
     m_endTime = microseconds;
 
     m_timeInMicro = m_endTime - m_startTime;
